feat: Post와 Comment간 순환 참조 발생 → @JsonIgnore를 이용해 순환 참조 해결, PostDto를 도입하여 엔터티가 외부에 노출되지 않도록 설계

+++ src/main/java/com/back/global/entity/BaseEntity.java
@@ -4,11 +4,12 @@ import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
 
 @MappedSuperclass
 @Getter
@@ -17,7 +18,7 @@ public abstract class BaseEntity {
 
     @Id
     @GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @CreatedDate
     private LocalDateTime createDate;
